+++ packages/service-apollo/src/index.js
@@ -17,29 +17,13 @@ module.exports = ({ registerAction, setContext, getConfig, registerHook }) => {
   });
 
   registerAction({
-    hook: '$FASTIFY_HACKS_BEFORE?',
+    hook: '$FASTIFY_PLUGIN?',
     name: SERVICE_NAME,
     trace: __filename,
-    handler: ({ fastify }, { getContext }) => {
-      const apollo = getContext('apollo', null);
-
-      // Prepare the shape of the decorators
-      fastify.decorate('apollo', apollo);
-      fastify.decorateRequest('apollo', null);
-      fastify.decorateReply('apollo', null);
-
-      // Add the references using hooks to comply with the decoratos API
-      // https://www.fastify.io/docs/v3.15.x/Decorators/
-
-      fastify.addHook('onRequest', (request, reply, done) => {
-        request.apollo = apollo;
-        done();
-      });
-
-      fastify.addHook('onResponse', (request, reply, done) => {
-        reply.apollo = apollo;
-        done();
-      });
+    handler: ({ decorate, decorateRequest }, { getContext }) => {
+      const apollo = getContext('apollo');
+      decorate('apollo', apollo);
+      decorateRequest('apollo', apollo);
     },
   });
 };