+++ packages/service-env/src/index.js
@@ -21,29 +21,13 @@ module.exports = ({ registerAction }) => {
 
   // Fastify Integration (optional hook)
   registerAction({
-    hook: '$FASTIFY_HACKS_AFTER?',
+    hook: '$FASTIFY_PLUGIN?',
     name: SERVICE_NAME,
     trace: __filename,
-    handler: ({ fastify }, { getContext }) => {
+    handler: ({ decorate, decorateRequest }, { getContext }) => {
       const getEnv = getContext('getEnv');
-
-      // Prepare the shape of the decorators
-      fastify.decorate('getEnv', getEnv);
-      fastify.decorateRequest('getEnv', null);
-      fastify.decorateReply('getEnv', null);
-
-      // Add the references using hooks to comply with the decoratos API
-      // https://www.fastify.io/docs/v3.15.x/Decorators/
-
-      fastify.addHook('onRequest', (request, reply, done) => {
-        request.getEnv = getEnv;
-        done();
-      });
-
-      fastify.addHook('onResponse', (request, reply, done) => {
-        reply.getEnv = getEnv;
-        done();
-      });
+      decorate('getEnv', getEnv);
+      decorateRequest('getEnv', getEnv);
     },
   });
 };