+++ packages/service-fastify/src/index.js
@@ -1,6 +1,5 @@
 const { INIT_SERVICE, START_SERVICE } = require('@forrestjs/hooks');
 const hooks = require('./hooks');
-const tddHandler = require('./tdd-handler');
 const initServiceHandler = require('./init-service-handler');
 const startServiceHandler = require('./start-service-handler');
 
@@ -12,7 +11,11 @@ const service = {
 module.exports = ({ registerAction, registerHook }) => {
   registerHook(hooks);
 
+  // The TDD support is strictly scoped to the development
+  // and test environment, even the module is conditionally
+  // loaded to minimize the memory footprint in production
   if (['development', 'test'].includes(process.env.NODE_ENV)) {
+    const tddHandler = require('./tdd-handler');
     registerAction({
       ...service,
       hook: '$FASTIFY_ROUTE',