+++ packages/service-fastify/src/init-service-handler.js
@@ -1,4 +1,5 @@
 const fastify = require('fastify');
+const axios = require('axios');
 const hooks = require('./hooks');
 
 module.exports = ({ getConfig, setContext, createHook, getContext }) => {
@@ -9,33 +10,39 @@ module.exports = ({ getConfig, setContext, createHook, getContext }) => {
   );
   const server = fastify(options);
 
+  // Register utilities to pass down to hooks:
   const registerPlugin = (...options) => server.register(...options);
   const decorate = (...options) => server.decorate(...options);
-  const decorateRequest = (...options) => server.decorateRequest(...options);
-  const decorateReply = (...options) => server.decorateReply(...options);
 
+  // Add the references using hooks to comply with the decoratos API
+  // https://www.fastify.io/docs/v3.15.x/Decorators/
+  const decorateRequest = (name, value) => {
+    server.decorateRequest(name, null);
+    server.addHook('onRequest', (request, reply, done) => {
+      request[name] = value;
+      done();
+    });
+  };
+  const decorateReply = (name, value) => {
+    server.decorateReply(name, null);
+    server.addHook('onResponse', (request, reply, done) => {
+      reply[name] = value;
+      done();
+    });
+  };
+
+  // Decorate the Fastify App with the basic tools
   server.decorate('getConfig', getConfig);
   server.decorate('getContext', getContext);
+  server.decorate('axios', axios);
 
-  // Prepare the shape of request/reply object
-  server.decorateRequest('getConfig', null);
-  server.decorateRequest('getContext', null);
-  server.decorateReply('getConfig', null);
-  server.decorateReply('getContext', null);
+  decorateRequest('getConfig', getConfig);
+  decorateRequest('getContext', getContext);
+  decorateRequest('axios', axios);
 
-  // Inject references in each request object
-  server.addHook('onRequest', (request, reply, done) => {
-    request.getConfig = getConfig;
-    request.getContext = getContext;
-    done();
-  });
-
-  // Inject references in each reply object
-  server.addHook('onResponse', (request, reply, done) => {
-    request.getConfig = getConfig;
-    request.getContext = getContext;
-    done();
-  });
+  decorateReply('getConfig', getConfig);
+  decorateReply('getContext', getContext);
+  decorateReply('axios', axios);
 
   createHook.sync(hooks.FASTIFY_HACKS_BEFORE, { fastify: server });
   createHook.sync(hooks.FASTIFY_PLUGIN, {