+++ packages/service-fastify/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Added globals utilities for E2E testing in Jest
 - Integrates with `service-env` to provide direct access to `getEnv` in the route handlers
 - Add `FASTIFY_OPTIONS` hook that lets hijack into the `fastify.instance.options` object at run time
+- `decoreateReuest` and `decorateReply` add the relative hooks
+  automatically
+- Add Moxios for stubbing HTTP requests made with Axios from the
+  test environment
 
 ## v1.13.1